feat(Modal, ModalContent): add support for title and sizes
affects: @entur/modal

This change introduces the concept of sizing to modals. Five sizes are available - from extraSmall
to extraLarge.

In addition, this change adds support for a standardized title, which all modals is
required to have.

BREAKING CHANGE:
The title prop is now required. Remove any top level heading, and use the title prop instead. The
size prop is now required, too.