feat(js tokens): add rem version for all relevant js tokens
affects: @entur/tokens

In addition to the px unit for tokens, it is now also possible to get the rem unit when in js.
Access then by adding a .rem after the token name like so: space.rem.large.
This could be usefull
when using styled components in tandem with stylesheets to avoid mixing units