chore: remove pre-push hook
The hook does not provide much security
because it's easy to skip it using
`git push --no-verify`.
Make it easier for people to contribute by removing it.

We can be confident in the code that gets pushed
because the same steps are already run in CI.
See .circleci/config.yml, jobs.build-test-and-deploy-to-preview.steps
at the step named "linter and test".