feat(radio panel): rewrite radioPanel component and add checkboxPanel component
affects: @entur/form

This change was done to make it clearer to the user if they are dealing with a checkbox or radio
button functionality when using radioPanel. Previously, radioPanel had a radio button functionality
but a checkbox visual icon. Now, a radio button is used for the visuals and checkboxPanel is added
to provide checkbox functionality

BREAKING CHANGE:
- "hideCheckbox" -> "hideRadioButton" in RadioPanel |
- Complete rewrite may lead to breakting changes |
- Visual changes