feat(searchable dropdown): new searchable dropdown component
affects: @entur/dropdown

complete rewrite of the searchable dropdown component. It is now available as a separate component
instead of a prop in the dropdown component. Should be more usable and stable.

BREAKING CHANGE:
is now a separate component instead of a prop in dropdown. | - is now only available asa controlled
component. 'value' is changed to 'selectedItem' and 'onChange' is required.